+++ src/styles/components/_menu.scss
@@ -7,3 +7,7 @@
   color: #EA859E;
 }
 
+#menu-item-form {
+  width: 75%;
+  margin: auto;
+}
